+++ agent/agent.go
@@ -149,7 +149,7 @@ func (a *Agent) buildTransportHeaderMap() http.Header {
 	header := http.Header{}
 	header.Set(transport.HeaderKeyNamespace, a.config.Namespace)
 	header.Set(transport.HeaderKeyAgentName, a.config.AgentName)
-	if tls := a.config.TLS; tls != nil && len(tls.CertFile) == 0 && len(tls.KeyFile) == 0 {
+	if tls := a.config.TLS; tls == nil || len(tls.CertFile) == 0 && len(tls.KeyFile) == 0 {
 		logger.Info("using password auth")
 		header.Set(transport.HeaderKeyUser, a.config.User)
 		userCredentials := fmt.Sprintf("%s:%s", a.config.User, a.config.Password)
@@ -384,7 +384,7 @@ func (a *Agent) StartStatsd(ctx context.Context) {
 	logger.Info("starting statsd server on address: ", a.statsdServer.MetricsAddr)
 
 	go func() {
-		if err := a.statsdServer.Run(ctx); err != nil && err != ctx.Err() {
+		if err := a.statsdServer.Run(ctx); err != nil && err != context.Canceled {
 			logger.WithError(err).Errorf("error with statsd server on address: %s, statsd listener will not run", a.statsdServer.MetricsAddr)
 		}
 	}()