+++ backend/agentd/agentd.go
@@ -48,16 +48,19 @@ type Agentd struct {
 	bus        messaging.MessageBus
 	tls        *corev2.TLSOptions
 	ringPool   *ringv2.Pool
+	ctx        context.Context
+	cancel     context.CancelFunc
 }
 
 // Config configures an Agentd.
 type Config struct {
-	Host     string
-	Port     int
-	Bus      messaging.MessageBus
-	Store    store.Store
-	TLS      *corev2.TLSOptions
-	RingPool *ringv2.Pool
+	Host         string
+	Port         int
+	Bus          messaging.MessageBus
+	Store        store.Store
+	TLS          *corev2.TLSOptions
+	RingPool     *ringv2.Pool
+	WriteTimeout int
 }
 
 // Option is a functional option.
@@ -65,6 +68,7 @@ type Option func(*Agentd) error
 
 // New creates a new Agentd.
 func New(c Config, opts ...Option) (*Agentd, error) {
+	ctx, cancel := context.WithCancel(context.Background())
 	a := &Agentd{
 		Host:     c.Host,
 		Port:     c.Port,
@@ -76,6 +80,8 @@ func New(c Config, opts ...Option) (*Agentd, error) {
 		wg:       &sync.WaitGroup{},
 		errChan:  make(chan error, 1),
 		ringPool: c.RingPool,
+		ctx:      ctx,
+		cancel:   cancel,
 	}
 
 	// prepare server TLS config
@@ -100,7 +106,7 @@ func New(c Config, opts ...Option) (*Agentd, error) {
 	a.httpServer = &http.Server{
 		Addr:         fmt.Sprintf("%s:%d", a.Host, a.Port),
 		Handler:      router,
-		WriteTimeout: 15 * time.Second,
+		WriteTimeout: time.Duration(c.WriteTimeout) * time.Second,
 		ReadTimeout:  15 * time.Second,
 		TLSConfig:    tlsServerConfig,
 		// Capture the log entries from agentd's HTTP server
@@ -140,6 +146,7 @@ func (a *Agentd) Start() error {
 
 // Stop Agentd.
 func (a *Agentd) Stop() error {
+	a.cancel()
 	if err := a.httpServer.Shutdown(context.TODO()); err != nil {
 		// failure/timeout shutting down the server gracefully
 		logger.Error("failed to shutdown http server gracefully - forcing shutdown")
@@ -207,7 +214,7 @@ func (a *Agentd) webSocketHandler(w http.ResponseWriter, r *http.Request) {
 
 	cfg.Subscriptions = addEntitySubscription(cfg.AgentName, cfg.Subscriptions)
 
-	session, err := NewSession(cfg, transport.NewTransport(conn), a.bus, a.store, unmarshal, marshal)
+	session, err := NewSession(a.ctx, cfg, transport.NewTransport(conn), a.bus, a.store, unmarshal, marshal)
 	if err != nil {
 		logger.WithError(err).Error("failed to create session")
 		http.Error(w, err.Error(), http.StatusInternalServerError)