+++ backend/agentd/session_test.go
@@ -2,14 +2,17 @@ package agentd
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"net/http"
 	"testing"
+	"time"
 
 	"github.com/gogo/protobuf/proto"
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/backend/messaging"
 	"github.com/sensu/sensu-go/testing/mockstore"
+	"github.com/sensu/sensu-go/testing/mocktransport"
 	"github.com/sensu/sensu-go/transport"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -74,7 +77,7 @@ func TestGoodSessionConfig(t *testing.T) {
 		Namespace:     "acme",
 		Subscriptions: []string{"testing"},
 	}
-	session, err := NewSession(cfg, conn, bus, st, UnmarshalJSON, MarshalJSON)
+	session, err := NewSession(context.Background(), cfg, conn, bus, st, UnmarshalJSON, MarshalJSON)
 	assert.NotNil(t, session)
 	assert.NoError(t, err)
 }
@@ -100,7 +103,7 @@ func TestGoodSessionConfigProto(t *testing.T) {
 		Namespace:     "acme",
 		Subscriptions: []string{"testing"},
 	}
-	session, err := NewSession(cfg, conn, bus, st, proto.Unmarshal, proto.Marshal)
+	session, err := NewSession(context.Background(), cfg, conn, bus, st, proto.Unmarshal, proto.Marshal)
 	assert.NotNil(t, session)
 	assert.NoError(t, err)
 }
@@ -129,7 +132,61 @@ func TestBadSessionConfig(t *testing.T) {
 	cfg := SessionConfig{
 		Subscriptions: []string{"testing"},
 	}
-	session, err := NewSession(cfg, conn, bus, st, UnmarshalJSON, MarshalJSON)
+	session, err := NewSession(context.Background(), cfg, conn, bus, st, UnmarshalJSON, MarshalJSON)
 	assert.Nil(t, session)
 	assert.Error(t, err)
+}
+
+func TestSessionTerminateOnSendError(t *testing.T) {
+	conn := new(mocktransport.MockTransport)
+	event := corev2.FixtureEvent("acme", "testing")
+	b, err := json.Marshal(event)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	tm := &transport.Message{
+		Payload: b,
+		Type:    transport.MessageTypeEvent,
+	}
+
+	conn.On("Receive").After(100*time.Millisecond).Return(tm, nil)
+	conn.On("Send", mock.Anything).Return(transport.ConnectionError{Message: "some horrible network outage"})
+	conn.On("Close").Return(nil)
+
+	bus, err := messaging.NewWizardBus(messaging.WizardBusConfig{})
+	if err != nil {
+		t.Fatal(err)
+	}
+	if err := bus.Start(); err != nil {
+		t.Fatal(err)
+	}
+
+	st := &mockstore.MockStore{}
+	st.On(
+		"GetNamespace",
+		mock.Anything,
+		"acme",
+	).Return(&corev2.Namespace{}, nil)
+
+	cfg := SessionConfig{
+		AgentName:     "testing",
+		Namespace:     "acme",
+		Subscriptions: []string{"testing"},
+	}
+	session, err := NewSession(context.Background(), cfg, conn, bus, st, UnmarshalJSON, MarshalJSON)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if err := session.Start(); err != nil {
+		t.Fatal(err)
+	}
+	if err := bus.Publish(messaging.SubscriptionTopic("acme", "testing"), corev2.FixtureCheckRequest("foo")); err != nil {
+		t.Fatal(err)
+	}
+	select {
+	case <-session.ctx.Done():
+	case <-time.After(time.Second * 5):
+		t.Fatal("broken session never stopped")
+	}
 }